refactor: update directory structure of project
