+++ apps/web/src/services/administration/type.ts
@@ -2,7 +2,7 @@
 import type { TimeStamp } from '@/schema/common/model';
 
 import type { RoleData } from '@/services/administration/iam/role/type';
-import type { ProjectGroupInfo, ProjectModel } from '@/services/project/type';
+import type { ProjectGroupInfo, ProjectModel } from '@/services/project/types/type';
 
 
 export interface RoleBindingType {
