+++ apps/web/src/services/alert-manager/escalation-policy/modules/EscalationPolicyProjectTree.vue
@@ -46,7 +46,7 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import type { ProjectGroup } from '@/services/asset-inventory/service-account/type';
-import type { ProjectGroupTreeItem, ProjectTreeRoot } from '@/services/project/type';
+import type { ProjectGroupTreeItem, ProjectTreeRoot } from '@/services/project/types/type';
 
 export default {
     name: 'EscalationPolicyProjectTree',
