+++ apps/web/src/services/project/modules/ProjectTree.vue
@@ -128,10 +128,10 @@ import SidebarTitle from '@/common/components/titles/sidebar-title/SidebarTitle.
 import ErrorHandler from '@/common/composables/error/errorHandler';
 import FavoriteButton from '@/common/modules/favorites/favorite-button/FavoriteButton.vue';
 
-import { useProjectPageStore } from '@/services/project/store/project-page-store';
+import { useProjectPageStore } from '@/services/project/stores/project-page-store';
 import type {
     ProjectItemResp,
-} from '@/services/project/type';
+} from '@/services/project/types/type';
 
 export default {
     name: 'ProjectTree',
