+++ apps/web/src/services/project/pages/ProjectAlertEventRulePage.vue
