+++ apps/web/src/services/project/pages/ProjectMaintenanceWindowPage.vue
@@ -107,7 +107,7 @@ import { useManagePermissionState } from '@/common/composables/page-manage-permi
 
 import MaintenanceWindowFormModal from '@/services/project/project-detail/modules/MaintenanceWindowFormModal.vue';
 import { MAINTENANCE_WINDOW_STATE } from '@/services/project/project-detail/project-alert/project-maintenance-window/lib/config';
-import { PROJECT_ROUTE } from '@/services/project/route-config';
+import { PROJECT_ROUTE } from '@/services/project/routes/route-config';
 
 const keyItemSets: KeyItemSet[] = [
     {
