+++ apps/web/src/services/project/pages/ProjectSummaryPage.vue
