+++ apps/web/src/services/project/project-detail/project-member/type.ts
@@ -4,7 +4,7 @@ import type { RawPagePermission } from '@/lib/access-control/config';
 import type { RouteQueryString } from '@/lib/router-query-string';
 
 import type { RoleData } from '@/services/administration/iam/role/type';
-import type { ProjectGroupInfo, ProjectModel } from '@/services/project/type';
+import type { ProjectGroupInfo, ProjectModel } from '@/services/project/types/type';
 
 export type ProjectMemberPageUrlQuery = Partial<Record<'filters', RouteQueryString>>;
