+++ apps/web/src/services/project/stores/project-detail-page-store.ts
